backend/requirements.txt
autopep8==1.5.6
isort==5.8.0
fastapi==0.63.0
fastapi-pagination==0.7.4
kubernetes==12.0.1
pydantic==1.8.1
python-dotenv==0.17.0
SQLAlchemy==1.4.11
starlette==0.13.6
uvicorn==0.13.4
pylint==2.8.2
python-jose==3.2.0
python-keycloak==0.24.0
psycopg2-binary==2.8.6
sqlalchemy-json==0.4.0